Fix the Linux/macOS build: sign-compare in the roundtrip self-test
int vs size_t comparison rejected by GCC/Clang -Werror; MSVC /W3 does
not flag it. Dev tool only; the shipped exe is untouched.

+++ src/roundtrip/main.cpp
@@ -427,7 +427,7 @@ namespace
 			return ExitHarness;
 		}
 
-		int proven = 0;
+		size_t proven = 0;
 		bool allOk = true;
 
 		// The two halves of the round-trip contract, on one saved source span: (a)
